Remove SyntaxNodeChildIterator.
