+++ crates/cairo-lang-syntax/src/node/iter.rs
@@ -1,102 +1,8 @@
-use std::collections::hash_map::Entry;
-use std::collections::HashMap;
 use std::ops::Deref;
-use std::sync::Arc;
 use std::vec;
 
-use cairo_lang_filesystem::span::TextOffset;
-use itertools::Itertools;
-
 use crate::node::db::SyntaxGroup;
-use crate::node::ids::GreenId;
-use crate::node::key_fields::get_key_fields;
-use crate::node::kind::SyntaxKind;
-use crate::node::stable_ptr::SyntaxStablePtr;
-use crate::node::{SyntaxNode, SyntaxNodeInner};
-
-pub struct SyntaxNodeChildIterator<'db> {
-    db: &'db dyn SyntaxGroup,
-    node: SyntaxNode,
-    green_iterator: vec::IntoIter<GreenId>,
-    /// The current offset in the source file of the start of the child.
-    offset: TextOffset,
-    /// Mapping from (kind, key_fields) to the number of times this indexing pair has been seen.
-    /// This is used to maintain the correct index for creating each StablePtr.
-    /// See [`get_key_fields`].
-    key_map: HashMap<(SyntaxKind, Vec<GreenId>), usize>,
-}
-
-impl<'db> Iterator for SyntaxNodeChildIterator<'db> {
-    type Item = SyntaxNode;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let green_id = self.green_iterator.next()?;
-        self.next_inner(green_id)
-    }
-}
-
-impl<'db> DoubleEndedIterator for SyntaxNodeChildIterator<'db> {
-    fn next_back(&mut self) -> Option<<SyntaxNodeChildIterator<'db> as Iterator>::Item> {
-        let green_id = self.green_iterator.next_back()?;
-        self.next_inner(green_id)
-    }
-}
-
-impl<'db> ExactSizeIterator for SyntaxNodeChildIterator<'db> {
-    fn len(&self) -> usize {
-        self.green_iterator.len()
-    }
-}
-
-impl<'db> SyntaxNodeChildIterator<'db> {
-    pub(super) fn new(node: &SyntaxNode, db: &'db dyn SyntaxGroup) -> Self {
-        SyntaxNodeChildIterator {
-            db,
-            node: node.clone(),
-            green_iterator: node
-                .green_node(db)
-                .children()
-                .iter()
-                .cloned()
-                .collect_vec()
-                .into_iter(),
-            offset: node.offset(),
-            key_map: HashMap::new(),
-        }
-    }
-
-    fn next_inner(
-        &mut self,
-        green_id: GreenId,
-    ) -> Option<<SyntaxNodeChildIterator<'db> as Iterator>::Item> {
-        let green = self.db.lookup_intern_green(green_id);
-        let width = green.width();
-        let kind = green.kind;
-        let key_fields: Vec<GreenId> = get_key_fields(kind, green.children());
-        let index = match self.key_map.entry((kind, key_fields.clone())) {
-            Entry::Occupied(mut entry) => entry.insert(entry.get() + 1),
-            Entry::Vacant(entry) => {
-                entry.insert(1);
-                0
-            }
-        };
-        let stable_ptr = self.db.intern_stable_ptr(SyntaxStablePtr::Child {
-            parent: self.node.0.stable_ptr,
-            kind,
-            key_fields,
-            index,
-        });
-        // Create the SyntaxNode view for the child.
-        let res = SyntaxNode(Arc::new(SyntaxNodeInner {
-            green: green_id,
-            offset: self.offset,
-            parent: Some(self.node.clone()),
-            stable_ptr,
-        }));
-        self.offset = self.offset.add_width(width);
-        Some(res)
-    }
-}
+use crate::node::SyntaxNode;
 
 /// `WalkEvent` describes tree walking process.
 #[derive(Debug, Copy, Clone)]
